cmd: move the conversion tool under cmd/meta
